backend/requirements.txt
torch
torchvision
pillow
numpy
typing
scikit-image
huggingface_hub
fastapi
python-multipart
uvicorn[standard]